Rename _slater.py to slater.py
It shouldn't be private
