+++ bfit/slater.py
